291 : creating media and mediaFolder services

+++ Rest/Controller/MediaFolderController.php
@@ -29,7 +29,7 @@ class MediaFolderController extends ARestController
      * Get collection of media folder
      * @return Symfony\Component\HttpFoundation\Response
      *
-     * @Rest\Pagination(default_count=25, max_count=100)
+     * @Rest\Pagination(default_count=100, max_count=200)
      *
      * @Rest\ParamConverter(
      *   name="parent", id_name="parent_uid", id_source="query", class="BackBee\NestedNode\MediaFolder", required=false
@@ -38,13 +38,10 @@ class MediaFolderController extends ARestController
      */
     public function getCollectionAction(Request $request, $start, $count, MediaFolder $parent = null)
     {
-        $qb = $this->getMediaFolderRepository()->createQueryBuilder("mf");
-        $qb->andParentIs($parent);
-        $paginator = new Paginator($qb->setFirstResult($start)->setMaxResults($count));
-        $results = [];
-        foreach ($paginator as $mediaFolder) {
+        $results = $this->getMediaFolderRepository()->getMediaFolders($parent,array("field"=>"_leftnode", "dir"=>"asc"));
+        /*foreach ($paginator as $mediaFolder) {
             $results[] = $mediaFolder;
-        }
+        }*/
         $response = $this->createJsonResponse($results);
         return $this->addRangeToContent($response, $results, $start);
     }
@@ -86,8 +83,15 @@ class MediaFolderController extends ARestController
      */
     public function putAction(MediaFolder $mediaFolder, Request $request)
     {
+        $parent_id = $request->get("parent_uid", null);
+        if (null === $parent_id) {
+            $parent = $this->getMediaFolderRepository()->getRoot();
+        }else{
+            $parent = $this->getMediaFolderRepository()->find($parent_id);
+        }
+
         $title = trim($request->request->get('title'));
-        if ($this->mediaFolderAlreadyExists($title)) {
+        if ($this->mediaFolderAlreadyExists($title, $parent)) {
             throw new BadRequestHttpException("A MediaFolder named '" . $title . "' already exists.");
         }
         $mediaFolder->setTitle($request->request->get('title'));
@@ -105,8 +109,19 @@ class MediaFolderController extends ARestController
         if (true === $mediaFolder->isRoot()) {
             throw new BadRequestHttpException('Cannot remove the root node of the MediaFodler');
         }
-        $this->getMediaFolderRepository()->delete($mediaFolder);
-        return new Response("", 204);
+        $isEmpty = $this->getMediaRepository()->countMedias($mediaFolder);
+        if (!$isEmpty) {
+            $response = new Response("", 204);
+            $this->getMediaFolderRepository()->delete($mediaFolder);
+        } else {
+            $response = new Response("MediaFolder [".$mediaFolder->getTitle()."] is not empty", 500);
+        }
+        return $response;
+    }
+
+    private function getMediaRepository()
+    {
+        return $this->getEntityManager()->getRepository('BackBee\NestedNode\Media');
     }
 
     /**
@@ -133,14 +148,17 @@ class MediaFolderController extends ARestController
                 $mediaFolder = new MediaFolder();
                 $mediaFolder->setUrl($request->request->get('url', Utils\String::urlize($title)));
                 $mediaFolder->setTitle($title);
-                if (null !== $parent) {
-                    $mediaFolder->setParent($parent);
-                    $this->getMediaFolderRepository()->insertNodeAsLastChildOf($mediaFolder, $parent);
-                } else {
-                    $root = $this->getMediaFolderRepository()->getRoot();
-                    $this->getMediaFolderRepository()->insertNodeAsLastChildOf($mediaFolder, $root);
+                if (null === $parent) {
+                    $parent = $this->getMediaFolderRepository()->getRoot();
                 }
+                if ($this->mediaFolderAlreadyExists($title, $parent)) {
+                    throw new BadRequestHttpException("A MediaFolder named '" . $title . "' already exists. in [" . $parent->getTitle() . "]");
+                }
+
+                $mediaFolder->setParent($parent);
+                $this->getMediaFolderRepository()->insertNodeAsLastChildOf($mediaFolder, $parent);
             }
+
             $this->getEntityManager()->persist($mediaFolder);
             $this->getEntityManager()->flush();
             return $this->createJsonResponse(null, 201, [
@@ -171,6 +189,7 @@ class MediaFolderController extends ARestController
         } catch (InvalidOperationSyntaxException $e) {
             throw new BadRequestHttpException('operation invalid syntax: ' . $e->getMessage());
         }
+
         $this->patchSiblingAndParentOperation($mediaFolder, $operations);
         $this->getEntityManager()->flush();
         return $this->createJsonResponse(null, 204);
@@ -183,7 +202,6 @@ class MediaFolderController extends ARestController
 
     private function patchSiblingAndParentOperation(MediaFolder $mediaFolder, &$operations)
     {
-
         $sibling_operation = null;
         $parent_operation = null;
 
@@ -195,6 +213,7 @@ class MediaFolderController extends ARestController
                 $parent_operation = $op;
             }
         }
+
         if (null !== $sibling_operation || null !== $parent_operation) {
             if ($mediaFolder->isRoot()) {
                 throw new BadRequestHttpException('Cannot move root node of a site.');
@@ -228,7 +247,7 @@ class MediaFolderController extends ARestController
     private function mediaFolderAlreadyExists($title, MediaFolder $parent)
     {
         $folderExists = false;
-        $medialFolder = $this->getMediaFolderRepository()->findOneBy(array("_title" => $title));
+        $medialFolder = $this->getMediaFolderRepository()->findOneBy(array("_title" => trim($title), "_parent" => $parent));
         if ($medialFolder) {
             $folderExists = true;
         }
